java-redis-common/src/main/java/org/isheihei/redis/common/util/ByteUtil.java
package org.isheihei.redis.common.util;

import java.nio.ByteBuffer;

/**
 * @ClassName: ByteUtil
 * @Description: TODO
 * @Date: 2022/6/16 21:14
 * @Author: isheihei
 */
public class ByteUtil {
    private static ByteBuffer buffer = ByteBuffer.allocate(8);
    public static byte[] intToBytes(int n) {
        byte[] b = new byte[4];
        b[3] = (byte) (n & 0xff);
        b[2] = (byte) (n >> 8 & 0xff);
        b[1] = (byte) (n >> 16 & 0xff);
        b[0] = (byte) (n >> 24 & 0xff);
        return b;
    }

    public static int bytesToInt(byte[] b){
        int res = 0;
        for(int i=0;i<b.length;i++){
            res += (b[i] & 0xff) << ((3-i)*8);
        }
        return res;
    }


    public static byte[] longToBytes(long x) {
        buffer.putLong(0, x);
        return buffer.array();
    }

    public static long bytesToLong(byte[] bytes) {
        buffer.put(bytes, 0, bytes.length);
        buffer.flip();//need flip
        return buffer.getLong();
    }

}

